ML_MAP/requirements.txt
mysql_connector_repackaged
numpy
pandas
python-dotenv
Requests
scikit_learn
tensorflow
